+++ app/src/main/java/ru/practicum/android/diploma/di/FilerModule.kt
@@ -1,0 +1,37 @@
+package ru.practicum.android.diploma.di
+
+import org.koin.dsl.module
+import ru.practicum.android.diploma.domain.filter.FilterRepositoryCountryFlow
+import ru.practicum.android.diploma.domain.filter.FilterRepositoryIndustriesFlow
+import ru.practicum.android.diploma.domain.filter.FilterRepositoryRegionFlow
+import ru.practicum.android.diploma.domain.filter.FilterRepositorySalaryBooleanFlow
+import ru.practicum.android.diploma.domain.filter.FilterRepositorySalaryTextFlow
+import ru.practicum.android.diploma.domain.filter.impl.FilterRepositoryCountryFlowImpl
+import ru.practicum.android.diploma.domain.filter.impl.FilterRepositoryIndustriesFlowImpl
+import ru.practicum.android.diploma.domain.filter.impl.FilterRepositoryRegionFlowImpl
+import ru.practicum.android.diploma.domain.filter.impl.FilterRepositorySalaryBooleanFlowImpl
+import ru.practicum.android.diploma.domain.filter.impl.FilterRepositorySalaryTextFlowImpl
+
+val filterModule = module {
+
+    factory<FilterRepositoryCountryFlow> {
+        FilterRepositoryCountryFlowImpl(get())
+    }
+
+    factory<FilterRepositoryRegionFlow> {
+        FilterRepositoryRegionFlowImpl(get())
+    }
+
+
+    factory<FilterRepositoryIndustriesFlow> {
+        FilterRepositoryIndustriesFlowImpl(get())
+    }
+
+    factory<FilterRepositorySalaryTextFlow> {
+        FilterRepositorySalaryTextFlowImpl(get())
+    }
+
+    factory<FilterRepositorySalaryBooleanFlow> {
+        FilterRepositorySalaryBooleanFlowImpl(get())
+    }
+}
