+++ app/src/main/java/ru/practicum/android/diploma/domain/filter/FilterRepositoryRegionFlow.kt
@@ -1,0 +1,11 @@
+package ru.practicum.android.diploma.domain.filter
+
+import kotlinx.coroutines.flow.StateFlow
+import ru.practicum.android.diploma.domain.filter.datashared.RegionShared
+
+interface FilterRepositoryRegionFlow {
+
+    fun setRegionFlow(region: RegionShared?)
+
+    fun getRegionFlow(): StateFlow<RegionShared?>
+}
