+++ app/src/main/java/ru/practicum/android/diploma/domain/filter/FilterRepositorySalaryBooleanFlow.kt
@@ -1,0 +1,11 @@
+package ru.practicum.android.diploma.domain.filter
+
+import kotlinx.coroutines.flow.StateFlow
+import ru.practicum.android.diploma.domain.filter.datashared.SalaryBooleanShared
+
+interface FilterRepositorySalaryBooleanFlow {
+
+    fun setSalaryBooleanFlow(salary: SalaryBooleanShared?)
+
+    fun getSalaryBooleanFlow(): StateFlow<SalaryBooleanShared?>
+}
